+++ packages/serverless-workflow-language-service/src/channel/SwfLanguageService.ts
@@ -36,8 +36,9 @@ import { Specification } from "@severlessworkflow/sdk-typescript";
 import { SW_SPEC_WORKFLOW_SCHEMA } from "../schemas";
 import { getLanguageService, TextDocument } from "vscode-json-languageservice";
 import { doCustomValidation } from "./customValidations";
-import { matchNodeWithLocation, SwfJsonPath, SwfLsNode } from "./matchNodeWithLocation";
+import { matchNodeWithLocation } from "./matchNodeWithLocation";
 import { findNodesAtLocation } from "./findNodesAtLocation";
+import { SwfJsonPath, SwfLsNode } from "./types";
 
 export type SwfLanguageServiceConfig = {
   shouldConfigureServiceRegistries: () => boolean; //TODO: See https://issues.redhat.com/browse/KOGITO-7107
