+++ packages/serverless-workflow-language-service/src/channel/SwfYamlLanguageService.ts
@@ -26,7 +26,7 @@ import {
 } from "yaml-language-server-parser";
 import { CodeLens, CompletionItem, Position, Range } from "vscode-languageserver-types";
 import { SwfLanguageService, SwfLanguageServiceArgs } from "./SwfLanguageService";
-import { SwfLsNode } from "./matchNodeWithLocation";
+import { SwfLsNode } from "./types";
 import { FileLanguage } from "../api";
 
 export class SwfYamlLanguageService {
